requirements.txt
rdflib
requests
lxml.html
lxml.etree
sys